// frontend/gen/gomoneypb/accounts/v1/accounts_pb.ts
export class ListAccountsResponse extends Message<ListAccountsResponse> {
  /**
   * @generated from field: repeated gomoneypb.accounts.v1.ListAccountsResponse.AccountItem accounts = 1;
   */
  accounts: ListAccountsResponse_AccountItem[] = [];

  constructor(data?: PartialMessage<ListAccountsResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "gomoneypb.accounts.v1.ListAccountsResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "accounts", kind: "message", T: ListAccountsResponse_AccountItem, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ListAccountsResponse {
    return new ListAccountsResponse().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ListAccountsResponse {
    return new ListAccountsResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ListAccountsResponse {
    return new ListAccountsResponse().fromJsonString(jsonString, options);
  }

  static equals(a: ListAccountsResponse | PlainMessage<ListAccountsResponse> | undefined, b: ListAccountsResponse | PlainMessage<ListAccountsResponse> | undefined): boolean {
    return proto3.util.equals(ListAccountsResponse, a, b);
  }
}

/**
 * @generated from message gomoneypb.accounts.v1.ListAccountsResponse.AccountItem
 */
export class ListAccountsResponse_AccountItem extends Message<ListAccountsResponse_AccountItem> {
  /**
   * @generated from field: gomoneypb.v1.Account account = 1;
   */
  account?: Account;

  constructor(data?: PartialMessage<ListAccountsResponse_AccountItem>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "gomoneypb.accounts.v1.ListAccountsResponse.AccountItem";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "account", kind: "message", T: Account },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ListAccountsResponse_AccountItem {
    return new ListAccountsResponse_AccountItem().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ListAccountsResponse_AccountItem {
    return new ListAccountsResponse_AccountItem().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ListAccountsResponse_AccountItem {
    return new ListAccountsResponse_AccountItem().fromJsonString(jsonString, options);
  }

  static equals(a: ListAccountsResponse_AccountItem | PlainMessage<ListAccountsResponse_AccountItem> | undefined, b: ListAccountsResponse_AccountItem | PlainMessage<ListAccountsResponse_AccountItem> | undefined): boolean {
    return proto3.util.equals(ListAccountsResponse_AccountItem, a, b);
  }
}